# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CommUnity Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /smart-drive:
    get:
      operationId: listSmartDriveFiles
      tags: [smart-hub]
      summary: List files in Smart Drive for an organization
      parameters:
        - name: organizationId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SmartDriveFile"
  /smart-drive/upload:
    post:
      operationId: uploadSmartDriveFile
      tags: [smart-hub]
      summary: Upload a file to Smart Drive
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                organizationId:
                  type: integer
                file:
                  type: string
                  format: binary
              required: [organizationId, file]
      responses:
        "201":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmartDriveFile"
  /smart-drive/{id}:
    patch:
      operationId: updateSmartDriveFile
      tags: [smart-hub]
      summary: Update file visibility and status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isVisibleToVolunteers:
                  type: boolean
                status:
                  type: string
                  enum: [approved, pending, rejected]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmartDriveFile"
    delete:
      operationId: deleteSmartDriveFile
      tags: [smart-hub]
      summary: Delete a file from Smart Drive
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Need Attachments ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SmartDriveFile:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        fileName:
          type: string
        filePath:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        uploadedBy:
          type: integer
        createdAt:
          type: string
          format: date-time
        isVisibleToVolunteers:
          type: boolean
        status:
          type: string
          enum: [approved, pending, rejected]
      required: [id, organizationId, fileName, filePath, fileType, fileSize, uploadedBy, createdAt, isVisibleToVolunteers, status]
...